C-46: use BigDecimal to calculate. Boxing

+++ src/main/java/com/boxing/cart/unit/calculator/Calculator.java
@@ -2,6 +2,8 @@ package com.boxing.cart.unit.calculator;
 
 import com.boxing.cart.function.InputInformation;
 
+import java.math.BigDecimal;
+
 public interface Calculator {
-    public abstract double calculate(double totalPrice, InputInformation inputInformation);
+    public abstract BigDecimal calculate(BigDecimal totalPrice, InputInformation inputInformation);
 }
